Refactor Wall and Water actions hpp=>h/cpp
