+++ src/openrct2-ui/interface/ViewportInteraction.cpp
@@ -23,7 +23,7 @@
 #include <openrct2/actions/LargeSceneryRemoveAction.h>
 #include <openrct2/actions/ParkEntranceRemoveAction.h>
 #include <openrct2/actions/SmallSceneryRemoveAction.h>
-#include <openrct2/actions/WallRemoveAction.hpp>
+#include <openrct2/actions/WallRemoveAction.h>
 #include <openrct2/localisation/Localisation.h>
 #include <openrct2/ride/Ride.h>
 #include <openrct2/ride/RideData.h>
