+++ src/openrct2/actions/WallRemoveAction.cpp
@@ -7,7 +7,7 @@
  * OpenRCT2 is licensed under the GNU General Public License version 3.
  *****************************************************************************/
 
-#pragma once
+#include "WallRemoveAction.h"
 
 #include "../Cheats.h"
 #include "../OpenRCT2.h"
@@ -17,108 +17,93 @@
 #include "../management/Finance.h"
 #include "../world/Location.hpp"
 #include "../world/Wall.h"
-#include "GameAction.h"
 
-DEFINE_GAME_ACTION(WallRemoveAction, GAME_COMMAND_REMOVE_WALL, GameActions::Result)
+void WallRemoveAction::AcceptParameters(GameActionParameterVisitor& visitor)
 {
-private:
-    CoordsXYZD _loc;
+    visitor.Visit(_loc);
+}
 
-public:
-    WallRemoveAction() = default;
-    WallRemoveAction(const CoordsXYZD& loc)
-        : _loc(loc)
+void WallRemoveAction::Serialise(DataSerialiser& stream)
+{
+    GameAction::Serialise(stream);
+
+    stream << DS_TAG(_loc);
+}
+
+GameActions::Result::Ptr WallRemoveAction::Query() const
+{
+    GameActions::Result::Ptr res = std::make_unique<GameActions::Result>();
+    res->Cost = 0;
+    res->Expenditure = ExpenditureType::Landscaping;
+
+    if (!LocationValid(_loc))
     {
+        return std::make_unique<GameActions::Result>(
+            GameActions::Status::InvalidParameters, STR_CANT_REMOVE_THIS, STR_INVALID_SELECTION_OF_OBJECTS);
     }
 
-    void AcceptParameters(GameActionParameterVisitor & visitor) override
+    const bool isGhost = GetFlags() & GAME_COMMAND_FLAG_GHOST;
+    if (!isGhost && !(gScreenFlags & SCREEN_FLAGS_SCENARIO_EDITOR) && !gCheatsSandboxMode && !map_is_location_owned(_loc))
     {
-        visitor.Visit(_loc);
+        return std::make_unique<GameActions::Result>(
+            GameActions::Status::NotOwned, STR_CANT_REMOVE_THIS, STR_LAND_NOT_OWNED_BY_PARK);
     }
 
-    void Serialise(DataSerialiser & stream) override
+    TileElement* wallElement = GetFirstWallElementAt(_loc, isGhost);
+    if (wallElement == nullptr)
     {
-        GameAction::Serialise(stream);
-
-        stream << DS_TAG(_loc);
+        return std::make_unique<GameActions::Result>(
+            GameActions::Status::InvalidParameters, STR_CANT_REMOVE_THIS, STR_INVALID_SELECTION_OF_OBJECTS);
     }
 
-    GameActions::Result::Ptr Query() const override
+    res->Cost = 0;
+    return res;
+}
+
+GameActions::Result::Ptr WallRemoveAction::Execute() const
+{
+    GameActions::Result::Ptr res = std::make_unique<GameActions::Result>();
+    res->Cost = 0;
+    res->Expenditure = ExpenditureType::Landscaping;
+
+    const bool isGhost = GetFlags() & GAME_COMMAND_FLAG_GHOST;
+
+    TileElement* wallElement = GetFirstWallElementAt(_loc, isGhost);
+    if (wallElement == nullptr)
     {
-        GameActions::Result::Ptr res = std::make_unique<GameActions::Result>();
-        res->Cost = 0;
-        res->Expenditure = ExpenditureType::Landscaping;
-
-        if (!LocationValid(_loc))
-        {
-            return std::make_unique<GameActions::Result>(
-                GameActions::Status::InvalidParameters, STR_CANT_REMOVE_THIS, STR_INVALID_SELECTION_OF_OBJECTS);
-        }
-
-        const bool isGhost = GetFlags() & GAME_COMMAND_FLAG_GHOST;
-        if (!isGhost && !(gScreenFlags & SCREEN_FLAGS_SCENARIO_EDITOR) && !gCheatsSandboxMode && !map_is_location_owned(_loc))
-        {
-            return std::make_unique<GameActions::Result>(
-                GameActions::Status::NotOwned, STR_CANT_REMOVE_THIS, STR_LAND_NOT_OWNED_BY_PARK);
-        }
-
-        TileElement* wallElement = GetFirstWallElementAt(_loc, isGhost);
-        if (wallElement == nullptr)
-        {
-            return std::make_unique<GameActions::Result>(
-                GameActions::Status::InvalidParameters, STR_CANT_REMOVE_THIS, STR_INVALID_SELECTION_OF_OBJECTS);
-        }
-
-        res->Cost = 0;
-        return res;
+        return std::make_unique<GameActions::Result>(
+            GameActions::Status::InvalidParameters, STR_CANT_REMOVE_THIS, STR_INVALID_SELECTION_OF_OBJECTS);
     }
 
-    GameActions::Result::Ptr Execute() const override
-    {
-        GameActions::Result::Ptr res = std::make_unique<GameActions::Result>();
-        res->Cost = 0;
-        res->Expenditure = ExpenditureType::Landscaping;
+    res->Position.x = _loc.x + 16;
+    res->Position.y = _loc.y + 16;
+    res->Position.z = tile_element_height(res->Position);
 
-        const bool isGhost = GetFlags() & GAME_COMMAND_FLAG_GHOST;
+    tile_element_remove_banner_entry(wallElement);
+    map_invalidate_tile_zoom1({ _loc, wallElement->GetBaseZ(), (wallElement->GetBaseZ()) + 72 });
+    tile_element_remove(wallElement);
 
-        TileElement* wallElement = GetFirstWallElementAt(_loc, isGhost);
-        if (wallElement == nullptr)
-        {
-            return std::make_unique<GameActions::Result>(
-                GameActions::Status::InvalidParameters, STR_CANT_REMOVE_THIS, STR_INVALID_SELECTION_OF_OBJECTS);
-        }
+    return res;
+}
 
-        res->Position.x = _loc.x + 16;
-        res->Position.y = _loc.y + 16;
-        res->Position.z = tile_element_height(res->Position);
-
-        tile_element_remove_banner_entry(wallElement);
-        map_invalidate_tile_zoom1({ _loc, wallElement->GetBaseZ(), (wallElement->GetBaseZ()) + 72 });
-        tile_element_remove(wallElement);
-
-        return res;
-    }
-
-private:
-    TileElement* GetFirstWallElementAt(const CoordsXYZD& location, bool isGhost) const
-    {
-        TileElement* tileElement = map_get_first_element_at(location);
-        if (!tileElement)
-            return nullptr;
-
-        do
-        {
-            if (tileElement->GetType() != TILE_ELEMENT_TYPE_WALL)
-                continue;
-            if (tileElement->GetBaseZ() != location.z)
-                continue;
-            if (tileElement->GetDirection() != location.direction)
-                continue;
-            if (tileElement->IsGhost() != isGhost)
-                continue;
-
-            return tileElement;
-        } while (!(tileElement++)->IsLastForTile());
+TileElement* WallRemoveAction::GetFirstWallElementAt(const CoordsXYZD& location, bool isGhost) const
+{
+    TileElement* tileElement = map_get_first_element_at(location);
+    if (!tileElement)
         return nullptr;
-    }
-};
+
+    do
+    {
+        if (tileElement->GetType() != TILE_ELEMENT_TYPE_WALL)
+            continue;
+        if (tileElement->GetBaseZ() != location.z)
+            continue;
+        if (tileElement->GetDirection() != location.direction)
+            continue;
+        if (tileElement->IsGhost() != isGhost)
+            continue;
+
+        return tileElement;
+    } while (!(tileElement++)->IsLastForTile());
+    return nullptr;
+}
