+++ src/openrct2/actions/WaterRaiseAction.h
@@ -1,0 +1,38 @@
+/*****************************************************************************
+ * Copyright (c) 2014-2020 OpenRCT2 developers
+ *
+ * For a complete list of all authors, please refer to contributors.md
+ * Interested in contributing? Visit https://github.com/OpenRCT2/OpenRCT2
+ *
+ * OpenRCT2 is licensed under the GNU General Public License version 3.
+ *****************************************************************************/
+
+#pragma once
+
+#include "GameAction.h"
+
+DEFINE_GAME_ACTION(WaterRaiseAction, GAME_COMMAND_RAISE_WATER, GameActions::Result)
+{
+private:
+    MapRange _range;
+
+public:
+    WaterRaiseAction() = default;
+    WaterRaiseAction(MapRange range)
+        : _range(range)
+    {
+    }
+
+    uint16_t GetActionFlags() const override
+    {
+        return GameAction::GetActionFlags();
+    }
+
+    void Serialise(DataSerialiser & stream) override;
+    GameActions::Result::Ptr Query() const override;
+    GameActions::Result::Ptr Execute() const override;
+
+private:
+    GameActions::Result::Ptr QueryExecute(bool isExecuting) const;
+    uint16_t GetHighestHeight(MapRange validRange) const;
+};
