+++ src/openrct2/libopenrct2.vcxproj
@@ -127,12 +127,12 @@
     <ClInclude Include="actions\TrackPlaceAction.h" />
     <ClInclude Include="actions\TrackRemoveAction.h" />
     <ClInclude Include="actions\TrackSetBrakeSpeedAction.h" />
-    <ClInclude Include="actions\WallPlaceAction.hpp" />
-    <ClInclude Include="actions\WallRemoveAction.hpp" />
-    <ClInclude Include="actions\WallSetColourAction.hpp" />
-    <ClInclude Include="actions\WaterLowerAction.hpp" />
-    <ClInclude Include="actions\WaterRaiseAction.hpp" />
-    <ClInclude Include="actions\WaterSetHeightAction.hpp" />
+    <ClInclude Include="actions\WallPlaceAction.h" />
+    <ClInclude Include="actions\WallRemoveAction.h" />
+    <ClInclude Include="actions\WallSetColourAction.h" />
+    <ClInclude Include="actions\WaterLowerAction.h" />
+    <ClInclude Include="actions\WaterRaiseAction.h" />
+    <ClInclude Include="actions\WaterSetHeightAction.h" />
     <ClInclude Include="audio\audio.h" />
     <ClInclude Include="audio\AudioChannel.h" />
     <ClInclude Include="audio\AudioContext.h" />
