+++ src/openrct2/ride/TrackDesign.cpp
@@ -27,8 +27,8 @@
 #include "../actions/SmallSceneryRemoveAction.h"
 #include "../actions/TrackPlaceAction.h"
 #include "../actions/TrackRemoveAction.h"
-#include "../actions/WallPlaceAction.hpp"
-#include "../actions/WallRemoveAction.hpp"
+#include "../actions/WallPlaceAction.h"
+#include "../actions/WallRemoveAction.h"
 #include "../audio/audio.h"
 #include "../core/DataSerialiser.h"
 #include "../core/File.h"
