+++ src/openrct2/world/Map.cpp
@@ -23,8 +23,8 @@
 #include "../actions/LargeSceneryRemoveAction.h"
 #include "../actions/ParkEntranceRemoveAction.h"
 #include "../actions/SmallSceneryRemoveAction.h"
-#include "../actions/WallRemoveAction.hpp"
-#include "../actions/WaterSetHeightAction.hpp"
+#include "../actions/WallRemoveAction.h"
+#include "../actions/WaterSetHeightAction.h"
 #include "../audio/audio.h"
 #include "../config/Config.h"
 #include "../core/Guard.hpp"
